move Team page to separate component
